Move handler to service initializer
